requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==4.4.0
mysqlclient==2.2.0
Pillow==9.5.0
requests==2.31.0
drf-spectacular==0.26.3
drf-yasg==1.21.7
PyJWT==2.9.0
sqlparse==0.5.3
typing_extensions==4.13.2
beautifulsoup4
